🐞 fix(alert): update receivers. sync silence in db, make swagger
